Move functions and configuration out of hamyambot.py file and into separate files.
