+++ hamyam.conf
@@ -1,0 +1,17 @@
+{
+	"QRZ_USERNAME":"",
+	"QRZ_PASSWORD":"",
+	"DISCORD_TOKEN":"",
+	"GUILD_IDS":[
+
+	],
+	"QRZ_URL":"https://xmldata.qrz.com/xml/current/",
+	"QRZ_PROFILE":"https://www.qrz.com/db/",
+	"DMRID_URL":"https://www.radioid.net/api/dmr/user",
+	"GRIDSQUARE_URL":"https://www.karhukoti.com/maidenhead-grid-square-locator/",
+	"CONDITIONS_URL":"http://www.hamqsl.com/solar101vhfpic.php",
+	"MUF_URL":"http://www.hamqsl.com/solarmuf.php",
+	"BANDS_URL":"https://i.imgur.com/j18VSeB.png",
+	"OSM_URL":"https://www.openstreetmap.org/",
+	"VERSION_STRING":"HAMYAMDiscord v3.00-BETA4 01/17/2023"
+}
