+++ hamyam/__init__.py
@@ -1,0 +1,10 @@
+import hamyam.call_by_dmrid
+import hamyam.callsign_color
+import hamyam.config
+import hamyam.distance
+import hamyam.dmr_by_call
+import hamyam.escapeChars
+import hamyam.help
+import hamyam.lookup_call
+import hamyam.qrz_lookup
+import hamyam.ziptogrid
