+++ hamyam/escapeChars.py
@@ -1,0 +1,11 @@
+def escapeChars(txt):
+	message = txt
+	temp = message.replace('-', '\-')
+	temp = temp.replace('.', '\.')
+	temp = temp.replace('#', '\#')
+	temp = temp.replace('(', '\(')
+	temp = temp.replace('=', '\=')
+	temp = temp.replace(')', '\)')
+	temp = temp.replace('+', '\+')
+	temp = temp.replace('!', '\!')
+	return temp
